+++ ui/src/components/collections/schema/SchemaFieldRelation.svelte
@@ -6,7 +6,7 @@
     import ObjectSelect from "@/components/base/ObjectSelect.svelte";
     import CollectionUpsertPanel from "@/components/collections/CollectionUpsertPanel.svelte";
     import SchemaField from "@/components/collections/schema/SchemaField.svelte";
-    import { collections } from "@/stores/collections";
+    import { collections, activeCollection } from "@/stores/collections";
 
     export let field;
     export let key = "";
@@ -200,7 +200,20 @@
             </div>
             <div class="col-sm-6">
                 <Field class="form-field" name="schema.{key}.options.cascadeDelete" let:uniqueId>
-                    <label for={uniqueId}>Delete main record on relation delete</label>
+                    <label for={uniqueId}>
+                        <span class="txt">Cascade delete</span>
+                        <i
+                            class="ri-information-line link-hint"
+                            use:tooltip={{
+                                text: `Whether on ${
+                                    selectedColection?.name || "relation"
+                                } record deletion to delete also the ${
+                                    $activeCollection?.name || "field"
+                                } associated records.`,
+                                position: "top",
+                            }}
+                        />
+                    </label>
                     <ObjectSelect
                         id={uniqueId}
                         items={defaultOptions}