+++ ui/src/components/records/RecordFileThumb.svelte
@@ -9,39 +9,63 @@
 
     let previewPopup;
     let thumbUrl = "";
-    let originalUrl = ApiClient.getFileUrl(record, filename);
+    let originalUrl = "";
+    let token = "";
+    let isLoadingToken = false;
+
+    loadFileToken();
 
     $: type = CommonHelper.getFileType(filename);
 
     $: hasPreview = ["image", "audio", "video"].includes(type) || filename.endsWith(".pdf");
 
-    $: thumbUrl = originalUrl ? originalUrl + "?thumb=100x100" : "";
+    $: originalUrl = !isLoadingToken ? ApiClient.files.getUrl(record, filename, { token }) : "";
+
+    $: thumbUrl = !isLoadingToken
+        ? ApiClient.files.getUrl(record, filename, { thumb: "100x100", token: token })
+        : "";
+
+    async function loadFileToken() {
+        isLoadingToken = true;
+
+        try {
+            token = await ApiClient.getAdminFileToken(record.collectionId);
+        } catch (err) {
+            console.warn("File token failure:", err);
+        }
+
+        isLoadingToken = false;
+    }
 
     function onError() {
         thumbUrl = "";
     }
 </script>
 
-<a
-    class="thumb {size ? `thumb-${size}` : ''}"
-    href={originalUrl}
-    target="_blank"
-    rel="noreferrer"
-    title={(hasPreview ? "Preview" : "Download") + " " + filename}
-    on:click|stopPropagation={(e) => {
-        if (hasPreview) {
-            e.preventDefault();
-            previewPopup?.show(originalUrl);
-        }
-    }}
->
-    {#if type === "image"}
-        <img src={thumbUrl} alt={filename} title="Preview {filename}" on:error={onError} />
-    {:else if type === "video" || type === "audio"}
-        <i class="ri-video-line" />
-    {:else}
-        <i class="ri-file-3-line" />
-    {/if}
-</a>
+{#if isLoadingToken}
+    <div class="thumb {size ? `thumb-${size}` : ''}" />
+{:else}
+    <a
+        class="thumb {size ? `thumb-${size}` : ''}"
+        href={originalUrl}
+        target="_blank"
+        rel="noreferrer"
+        title={(hasPreview ? "Preview" : "Download") + " " + filename}
+        on:click|stopPropagation={(e) => {
+            if (hasPreview) {
+                e.preventDefault();
+                previewPopup?.show(originalUrl);
+            }
+        }}
+    >
+        {#if type === "image"}
+            <img src={thumbUrl} alt={filename} title="Preview {filename}" on:error={onError} />
+        {:else if type === "video" || type === "audio"}
+            <i class="ri-video-line" />
+        {:else}
+            <i class="ri-file-3-line" />
+        {/if}
+    </a>
+{/if}
 
 <PreviewPopup bind:this={previewPopup} />