+++ ui/src/utils/ApiClient.js
@@ -1,10 +1,15 @@
-import PocketBase, { LocalAuthStore, Admin } from "pocketbase";
+import PocketBase, { LocalAuthStore, Admin, isTokenExpired } from "pocketbase";
 // ---
-import CommonHelper      from "@/utils/CommonHelper";
-import { replace }       from "svelte-spa-router";
-import { addErrorToast } from "@/stores/toasts";
-import { setErrors }     from "@/stores/errors";
-import { setAdmin }      from "@/stores/admin";
+import CommonHelper                     from "@/utils/CommonHelper";
+import { replace }                      from "svelte-spa-router";
+import { get }                          from "svelte/store";
+import { addErrorToast }                from "@/stores/toasts";
+import { setErrors }                    from "@/stores/errors";
+import { setAdmin }                     from "@/stores/admin";
+import { protectedFilesCollectionsCache } from "@/stores/collections";
+
+
+const adminFileTokenKey = "pb_admin_file_token";
 
 /**
  * Clears the authorized state and redirects to the login page.
@@ -15,7 +20,7 @@ PocketBase.prototype.logout = function(redirect = true) {
     this.authStore.clear();
 
     if (redirect) {
-        replace('/login');
+        replace("/login");
     }
 };
 
@@ -26,7 +31,7 @@ PocketBase.prototype.logout = function(redirect = true) {
  * @param  {Boolean} notify     Whether to add a toast notification.
  * @param  {String}  defaultMsg Default toast notification message if the error doesn't have one.
  */
-PocketBase.prototype.errorResponseHandler = function(err, notify = true, defaultMsg = '') {
+PocketBase.prototype.errorResponseHandler = function(err, notify = true, defaultMsg = "") {
     if (!err || !(err instanceof Error) || err.isAbort) {
         return;
     }
@@ -59,9 +64,45 @@ PocketBase.prototype.errorResponseHandler = function(err, notify = true, default
     // forbidden
     if (statusCode === 403) {
         this.cancelAllRequests();
-        return replace('/');
+        return replace("/");
     }
 };
+
+/**
+ * @return {Promise<String>}
+ */
+PocketBase.prototype.getAdminFileToken = async function(collectionId = "") {
+    let needToken = true;
+
+    if (collectionId) {
+        const protectedCollections = get(protectedFilesCollectionsCache);
+        needToken = typeof protectedCollections[collectionId] !== "undefined"
+            ? protectedCollections[collectionId]
+            : true;
+    }
+
+    if (!needToken) {
+        return "";
+    }
+
+    let token = localStorage.getItem(adminFileTokenKey) || "";
+
+    // request a new token only if the previous one is missing or will expire soon
+    if (!token || isTokenExpired(token, 15)) {
+        // remove previously stored token (if any)
+        token && localStorage.removeItem(adminFileTokenKey);
+
+        if (!this._adminFileTokenRequest) {
+            this._adminFileTokenRequest = this.files.getToken();
+        }
+
+        token = await this._adminFileTokenRequest;
+        localStorage.setItem(adminFileTokenKey, token);
+        this._adminFileTokenRequest = null;
+    }
+
+    return token;
+}
 
 // Custom auth store to sync the svelte admin store state with the authorized admin instance.
 class AppAuthStore extends LocalAuthStore {